services/workflow-service/pyproject.toml
[project]
name = "workflow-service"
version = "0.1.0"
description = "Repository lifecycle and workflow orchestration"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlmodel>=0.0.22",
    "psycopg>=3.2.0",
    "pydantic-settings>=2.0.0",
    "pyjwt>=2.10.0",
    "cryptography>=43.0.0",
    "structlog>=24.0.0",
    "httpx>=0.27.0",
    "tenacity>=9.0.0",
    "boto3>=1.35.0",
]

[dependency-groups]
test = [
    "pytest>=8.3.0",
    "testcontainers[postgres]>=4.8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = [".", ".."]
addopts = "-v --tb=short -p no:warnings"

[tool.uv]
managed = true
